Cola de Tareas - Ahora el sistema puede asignar tareas a una cola. - Se creó un worker de email. Ahora al momento de hacer una compra, eliminar un producto o restablecer la contraseña se asigna la tarea a una cola. Fix - Se arregló un problema que no permitia ver las ventas de un vendedor si el comprador no tenia direccion asignada. - Se arregló la autenticación por tercero(Github). Ademas se agregó un DTO al momento de crear el usuario. - Se arregló un problema con las variables de entorno. Y se eliminó el archivo en desuso dotenv.helper.js Cambios - Se cambió los minutos de espera del rate limit de 15 a 5

+++ src/service/products.service.js
@@ -208,7 +208,7 @@ class ProductsService{
     }else {
 
         //Enviar mail al propietario del producto si el producto no es de un Admin
-        if (productFound.owner !== "Admin") {
+        if (productFound.owner !== "admin") {
           await sendEmailDeleteProduct(user.email, productFound);
         }
 
@@ -217,8 +217,23 @@ class ProductsService{
     
   }
 
-  async deleteFromSeller(id_product_seller){
-     return this.repository.deleteFromSeller(id_product_seller);
+  async deleteFromSeller(user, id_product_seller){
+
+    const productSeller = await this.repository.findBySellerProduct(user.id, id_product_seller); 
+
+    if(!productSeller){
+      throw new NotFoundException("El producto no existe o no tienes permiso para eliminar este producto");
+    }
+    
+    //eliminar producto del vendedor
+    const productDeleted = await this.repository.deleteFromSeller(id_product_seller);
+
+    //Enviar mail al propietario del producto si el producto no es de un Admin
+    if (user.rol !== "admin") {
+      await sendEmailDeleteProduct(user.email, productSeller);
+    }
+
+     return productDeleted;
   }
